pyproject.toml
[project]
name = "guthrieai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "python-dotenv>=1.2.1",
]
